add npm adduser to install command and refactor NpmFunctions.ts file, add pkg script to package.json
