+++ package.json
@@ -8,6 +8,16 @@
   "keywords": [
     "CLI"
   ],
+  "pkg": {
+    "scripts": [
+      "lib/**/*.js",
+      "lib/auth/*.js"
+    ],
+    "assets": [
+      "node_modules/npm/node_modules/node-gyp/**/*",
+      "node_modules/npm/lib/"
+    ]
+  },
   "files": [
     "lib"
   ],
