+++ packages/imperative/src/plugins/cmd/install/install.definition.ts
@@ -51,6 +51,14 @@ const fileDescription =
     "way to install plug-ins that were lost or corrupted after " +
     `reinstalling or updating ${ImperativeConfig.instance.loadedConfig.productDisplayName}.`;
 
+const loginDescription =
+    "The flag to add a registry user account to install from secure registry. It saves credentials " +
+    "to the .npmrc file using `npm adduser`. When this value is omitted, credentials from .npmrc file is used. " +
+    "If you used this flag once for specific registry, you don't have to use it again, it uses credentials from .npmrc file.\n" +
+    "\n" +
+    "For more information about npm registries, see: " +
+    "https://docs.npmjs.com/cli/adduser";
+
 /**
  * Definition of the install command.
  * @type {ICommandDefinition}
@@ -82,6 +90,13 @@ export const installDefinition: ICommandDefinition = {
             type: "string",
             description: registryDescription,
             required: false
+        },
+        {
+            name: "login",
+            type: "boolean",
+            description: loginDescription,
+            required: false,
+            implies: ["registry"]
         }
     ],
     examples: [
